Split backend and frontend
