+++ third_deliverable/frontend/go.mod
@@ -1,0 +1,3 @@
+module github.com/Ciip1996/academy-go-q12021
+
+go 1.16
