--- third_deliverable/go.mod
@@ -1,3 +1,0 @@
-module github.com/Ciip1996/academy-go-q12021
-
-go 1.16
